llama3/Hackathon/prompt-portal/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
SQLAlchemy==2.0.30
alembic==1.13.1
python-multipart==0.0.9
paho-mqtt==2.1.0
